+++ src/Router.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import BusinessListDetailPage from "./pages/ActivityReport/BusinessListDetailPage";
 import BusinessListPage from "./pages/ActivityReport/BusinessListPage";
 import BusinessSearchResultPage from "./pages/ActivityReport/BusinessSearchResultPage";
+import CoalitionDetailPage from "./pages/ActivityReport/CoalitionDetailPage";
 import CoalitionPage from "./pages/ActivityReport/CoalitionPage";
 import PolicyListPage from "./pages/ActivityReport/PolicyListPage";
 import Home from "./pages/Home";
@@ -26,7 +27,10 @@ const Router = () => {
           element={<BusinessSearchResultPage />}
         />
         <Route path="/activityReport/coalition" element={<CoalitionPage />} />
-
+        <Route
+          path="/activityReport/coalitionDetail/:id"
+          element={<CoalitionDetailPage />}
+        />
         {/* </Route> */}
       </Routes>
     </BrowserRouter>
