+++ src/pages/ActivityReport/CoalitionDetailPage.tsx
@@ -1,0 +1,11 @@
+import GoListButton from "@/components/common/GoListButton";
+
+const CoalitionDetailPage = () => {
+  return (
+    <div>
+      <GoListButton />
+    </div>
+  );
+};
+
+export default CoalitionDetailPage;
