+++ src/styles/ActivityReport/BusinessDetail/BusinessListDetailPageStyle.ts
@@ -5,19 +5,3 @@ export const Container = styled.div`
   width: 100vw;
   flex-direction: column;
 `;
-
-export const ButtonContainer = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: center;
-`;
-
-export const GoListButton = styled.button`
-  display: flex;
-  justify-content: center;
-  background-color: var(--Primary);
-  height: 40px;
-  width: 100px;
-  margin-top: 40px;
-  border-radius: 2px;
-`;
