+++ src/styles/common/GoListButtonStyle.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const ButtonContainer = styled.div`
+  display: flex;
+  width: 100%;
+  justify-content: center;
+`;
+
+export const GoListButton = styled.button`
+  display: flex;
+  justify-content: center;
+  background-color: var(--Primary);
+  height: 40px;
+  width: 100px;
+  margin-top: 40px;
+  border-radius: 2px;
+`;
